End card and implemented on projects
