+++ projects/components/src/lib/hyt-detailed-card/hyt-detailed-card.component.scss
@@ -1,14 +1,14 @@
-$primary: #0956B6;
+$primary: #0956b6;
 $primaryDarker: #094894;
-.hyt-project-card{
+.hyt-project-card {
     position: relative;
-    width: 100%;
-    height: 300px;
-    margin: 1rem 0;
-    
-    mat-card{
+    width: 225px;
+    aspect-ratio: 4/5;
+    margin: 1rem;
+
+    mat-card {
         background-color: #0d416d;
-        background-image : url('../../../../assets/images/projects/Card_Idle.png');
+        background-image: url("../../../../assets/images/projects/Card_Idle.png");
         background-repeat: no-repeat;
         background-position: bottom center;
         background-size: cover;
@@ -17,26 +17,38 @@ $primaryDarker: #094894;
         overflow: hidden;
         border-radius: 6px;
         color: #fff;
-        font-family: 'Exo', sans-serif;
+        font-family: "Exo", sans-serif;
 
-        .title-front-card{
+        .title-front-card {
             position: absolute;
             bottom: 0;
             left: 0;
             width: 100%;
-            font-size: 1.4em;
-            line-height: 26px;
-            padding: 6px;
-            background: linear-gradient(to right, $primary, 66% $primaryDarker);
+            font-size: 1.5em;
+            line-height: 1.5em;
+            padding: 8px;
+            background: rgb(9, 72, 148);
+            background: linear-gradient(0deg, $primaryDarker 0%, #2558a5c7 100%);
             overflow: hidden;
             display: -webkit-box;
             line-clamp: 2;
             word-wrap: break-word;
             -webkit-line-clamp: 2;
             -webkit-box-orient: vertical;
+
+            &.title-front-card-custom-bg {
+                background: #2558a5c7;
+                background: linear-gradient(0deg, #084185 0%, #2558a5c7 50%, #2558a5c7 100%);
+            }
         }
 
-        .box-card-hover{
+        .detailed-front-icon {
+            position: absolute;
+            right: 6px;
+            top: 8px;
+        }
+
+        .box-card-hover {
             position: absolute;
             top: 0;
             left: 0;
@@ -45,28 +57,28 @@ $primaryDarker: #094894;
             display: flex;
             justify-content: space-between;
             flex-direction: column;
-            padding: 6px;
+            padding: 6px 6px 44px;
             background-image: url("../../../../assets/images/projects/Card_Round-Hover.png");
             background-repeat: no-repeat;
             background-position: bottom center;
             background-size: cover;
-            transition: all .2s ease-in-out;
-            transform: translateY(100%) scale(0) ;
+            transition: all 0.15s ease-in;
+            transform: translateY(100%) scale(0);
             opacity: 0;
 
-            .content-card{
+            .content-card {
                 display: grid;
 
-                .card-title{
+                .card-title {
                     display: block;
                     white-space: nowrap;
                     overflow: hidden;
                     text-overflow: ellipsis;
                     font-size: 1.2em;
                     line-height: 35px;
-                    color: #0956B6;
+                    color: #0956b6;
                 }
-                .card-description{
+                .card-description {
                     color: #414141;
                     overflow: hidden;
                     display: -webkit-box;
@@ -80,22 +92,22 @@ $primaryDarker: #094894;
                     -webkit-box-orient: vertical;
                 }
             }
-            
-            .card-elements{
-                color: #0956B6;
-                margin-bottom: 8px;
 
-                .single-element{
+            .card-elements {
+                color: #0956b6;
+                margin-bottom: 12px;
+
+                .single-element {
                     display: flex;
                     align-items: center;
-                    
-                    mat-icon{
-                        font-size: 22px;
+
+                    .hyt-icon {
+                        font-size: 26px;
                         margin-right: 5px;
                     }
-                    span{
-                        font-size: 15px;
-                        line-height: 16px;
+                    span {
+                        font-size: 16px;
+                        line-height: 26px;
                     }
                 }
             }
@@ -104,6 +116,32 @@ $primaryDarker: #094894;
                 height: 40px;
                 font-size: 16px;
                 line-height: 16px;
+                display: flex;
+                position: absolute;
+                bottom: 6px;
+                left: 0;
+                width: 100%;
+                .detailed-primary-btn {
+                    width: 90%;
+                    margin: 0 5%;
+                }
+
+                &.btn-card-secondary {
+                    height: 36px;
+
+                    .detailed-primary-btn {
+                        width: 60%;
+                        margin: 0 2% 0 8%;
+                        font-size: 15px;
+                        line-height: 13px;
+                    }
+                    .detailed-secondary-btn {
+                        width: 20%;
+                        margin: 3px 8% 3px 2%;
+                        font-size: 16px;
+                        line-height: 16px;
+                    }
+                }
             }
         }
 
@@ -116,22 +154,40 @@ $primaryDarker: #094894;
         }
     }
 
-    @media screen and (max-width: 1600px) {
-        height: 250px;
+    .delete-project-btn {
+        width: 30px;
+        height: 30px;
+        background-color: #523dd6;
+        text-align: center;
+        position: absolute;
+        top: -16px;
+        right: -12px;
+        color: #fff;
+        border-radius: 50%;
+        padding-top: 5px;
+        cursor: pointer;
+        display: none;
+    }
+    
+    &:hover {
+        .delete-project-btn {
+            display: block;
+            &:hover {
+                background-color: #d40b0b;
+            }
+        }
     }
 }
 
-::ng-deep{
-    hyt-detailed-card{
-        width: 100%;
-    }
-}
-
-:host ::ng-deep{
-    .detailed-card-btn{
-        button{
+:host ::ng-deep {
+    .detailed-card-btn {
+        button {
             height: 100%;
             min-height: 100%;
+        }
+        &.detailed-secondary-btn button {
+            border-radius: 1rem;
+            max-width: 40px;
         }
     }
 }
