+++ projects/components/src/lib/hyt-filter-button/hyt-filter-button.component.html
@@ -1,4 +1,4 @@
-<div class="row d-flex flex-wrap align-items-center justify-content-center mt-3 mt-md-0">
+<div class="row d-flex flex-wrap align-items-center justify-content-center mt-2 mt-md-0">
     <span [innerHTML]="label" class="col-12 text-center mb-1"></span>
     <div class="hyt-filter-button-container d-flex flex-wrap align-items-center justify-content-center">
         <label *ngFor="let option of options" class="hyt-filter-button hyt-filter-button-{{shape}}">
