--- src/app/pages/projects/project-card/project-card.component.html
@@ -1,58 +1,0 @@
-
-<mat-card class="mat-elevation-z5">
-    
-    <div class="title-card-back"> <!-- flex-fill d-flex align-items-end -->
-        {{ project.name }}
-    </div>
-
-    <div class="box-card-front">
-
-        <div class="header-card-front">
-
-            {{ project.name }}
-
-        </div> <!-- End Card Front Header -->
-
-        <div class="content-card-front">
-
-            <div class="card-description">
-                {{ project.description }}
-            </div> <!-- End Card Description -->
-
-            <div class="card-elements">
-
-                <div class="single-element">
-                    <mat-icon>devices_other</mat-icon>
-                    <strong class="device-count px-2">{{ deviceCount }}</strong>
-                    <span i18n="@@HYT_projectcard_string_device">Device{{ deviceCount == 1 ? '' : 's' }}</span>
-                </div> <!-- End Single Element -->
-
-                <div class="single-element">
-                    <mat-icon>code</mat-icon>
-                    <strong class="count-element px-2">{{ rulesCount }}</strong> 
-                    <span i18n="@@HYT_projectcard_string_rule">Rule{{ rulesCount == 1 ? '' : 's' }}</span>
-                </div> <!-- End Single Element -->
-                
-
-            </div> <!-- End Card Elements -->
-
-        </div> <!-- End Card Front Content -->
-
-        <div class="footer-card-front">
-
-            <hyt-button
-                class="view-project-btn" 
-                routerLink="{{project.id}}"
-            >
-                View Project
-            </hyt-button>
-            
-        </div> <!-- End Card Front Footer -->
-
-    </div>
-
-</mat-card>
-
-<div class="delete-project-btn" (click)="deleteProject()">
-    <span class="hyt-icon icon-hyt_cancel"></span>
-</div>
