+++ src/styles.scss
@@ -1,6 +1,6 @@
 /* You can add global styles to this file, and also import other style files */
 @import "~@angular/material/prebuilt-themes/indigo-pink.css";
-@import './assets/styles/main';
+@import "./assets/styles/main";
 
 .cdk-overlay-pane .base-dialog {
     width: 100%;
@@ -8,13 +8,32 @@
 }
 
 #bodyMap {
-    .blue{fill-rule:evenodd;clip-rule:evenodd;fill:#144B64;}
-    .blueDark{fill:#001A24;}
-    .Glow{fill:#001A24;filter:url(#glow);}
-    .st0{filter:url(#glow);}
-    .st1{fill:#001A24;}
-    .st2{opacity:0.0;fill:none;}
-    .st3{opacity:0.6;fill:#EA0707;}
+    .blue {
+        fill-rule: evenodd;
+        clip-rule: evenodd;
+        fill: #144b64;
+    }
+    .blueDark {
+        fill: #001a24;
+    }
+    .Glow {
+        fill: #001a24;
+        filter: url(#glow);
+    }
+    .st0 {
+        filter: url(#glow);
+    }
+    .st1 {
+        fill: #001a24;
+    }
+    .st2 {
+        opacity: 0;
+        fill: none;
+    }
+    .st3 {
+        opacity: 0.6;
+        fill: #ea0707;
+    }
 }
 
 .toastAlarmUp {
@@ -33,4 +52,11 @@
     font-size: 0.8rem;
     background-color: #266dca;
 }
-  
+.hyt-card-container {
+    display: flex;
+    flex-wrap: wrap;
+    margin-bottom: 1rem;
+    @media screen and (max-width: 1153px) {
+        justify-content: center;
+    }
+}
